\name{+.AdditiveUnit}
\alias{+.AdditiveUnit}
\title{
Add heatmaps or row annotations to a heatmap list

}
\description{
Add heatmaps or row annotations to a heatmap list

}
\usage{
\method{+}{AdditiveUnit}(x, y)}
\arguments{

  \item{x}{a \code{\link{EnrichedHeatmap-class}} object, a \code{\link{Heatmap-class}} object, a \code{\link{HeatmapAnnotation-class}} object or a \code{\link{HeatmapList-class}} object.}
  \item{y}{a \code{\link{EnrichedHeatmap-class}} object, a \code{\link{Heatmap-class}} object, a \code{\link{HeatmapAnnotation-class}} object or a \code{\link{HeatmapList-class}} object.}
}
\details{
It overwrites \code{\link[ComplexHeatmap]{+.AdditiveUnit}}.

}
\value{
A \code{\link[ComplexHeatmap]{HeatmapList-class}} object or a \code{\link{EnrichedHeatmapList-class}} object

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
