\name{CentralizedHeatmap-class}
\docType{class}
\alias{CentralizedHeatmap-class}
\title{
Class for a single heatmap

}
\description{
Class for a single heatmap

}
\details{
The structure of \code{\link{CentralizedHeatmap-class}} is the same as
\code{\link[ComplexHeatmap]{HeatmapList-class}} and the class is inherited from \code{\link[ComplexHeatmap]{Heatmap-class}}.

The \code{\link{CentralizedHeatmap-class}} pre-defines some parameters for \code{\link[ComplexHeatmap]{Heatmap-class}} such as
the order of rows and supressing column clustering. Also there are several
new parameters that are attached in the object.

}
\section{Methods}{
The \code{\link{CentralizedHeatmap-class}} provides following methods:

\itemize{
  \item \code{\link{CentralizedHeatmap}}: constructor method.
  \item \code{\link{draw,CentralizedHeatmap-method}}: draw a single heatmap.
}

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
