\name{CentralizedHeatmapList-class}
\docType{class}
\alias{CentralizedHeatmapList-class}
\title{
Class for a list of heatmaps

}
\description{
Class for a list of heatmaps

}
\details{
The structure of \code{\link{CentralizedHeatmapList-class}} is the same as
\code{\link[ComplexHeatmap]{HeatmapList-class}} and the class is inherited from \code{\link[ComplexHeatmap]{HeatmapList-class}}.

}
\section{Methods}{
The \code{\link{CentralizedHeatmapList-class}} provides following methods:

\itemize{
  \item \code{\link{draw,CentralizedHeatmapList-method}}: draw a list of heatmaps.
}

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
