\name{CentralizedHeatmapList}
\alias{CentralizedHeatmapList}
\title{
Constructor method for CentralizedHeatmapList class

}
\description{
Constructor method for CentralizedHeatmapList class

}
\usage{
CentralizedHeatmapList(...)}
\arguments{

  \item{...}{arguments}
}
\details{
There is no public constructor method for the \code{\link{CentralizedHeatmapList-class}}.

}
\value{
No value is returned.

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
