\name{draw-dispatch}
\alias{draw}
\title{
Method dispatch page for draw}
\description{
Method dispatch page for \code{draw}.}
\section{Dispatch}{
\code{draw} can be dispatched on following classes:

\itemize{
\item \code{\link{draw,CentralizedHeatmapList-method}}, \code{\link{CentralizedHeatmapList-class}} class method
\item \code{\link{draw,CentralizedHeatmap-method}}, \code{\link{CentralizedHeatmap-class}} class method
}
}
