\name{EnrichedHeatmapList-class}
\docType{class}
\alias{EnrichedHeatmapList-class}
\title{
Class for a list of heatmaps
}
\description{
Class for a list of heatmaps
}
\details{
The \code{\link{EnrichedHeatmapList-class}} is inherited from \code{\link[ComplexHeatmap]{HeatmapList-class}}.
}
\section{Methods}{
The \code{\link{EnrichedHeatmapList-class}} provides following methods:

\itemize{
  \item \code{\link{draw,EnrichedHeatmapList-method}}: draw a list of heatmaps.
}}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
