\name{EnrichedHeatmapList}
\alias{EnrichedHeatmapList}
\title{
Constructor method for EnrichedHeatmapList class  


}
\description{
Constructor method for EnrichedHeatmapList class  


}
\usage{
EnrichedHeatmapList(...)
}
\arguments{

  \item{...}{arguments}

}
\details{
There is no public constructor method for the \code{\link{EnrichedHeatmapList-class}}.  


}
\value{
No value is returned.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\section{Example}{
# no example NULL 


}
