\name{failed_rows}
\alias{failed_rows}
\title{
Indices of Rows Failed from Smoothing
}
\description{
Indices of Rows Failed from Smoothing
}
\usage{
failed_rows(m)
}
\arguments{

  \item{m}{Matrix from \code{\link{normalizeToMatrix}}.}

}
\value{
A numeric vector or \code{NULL}.
}
\examples{
# There is no example
NULL
}
