\name{print.normalizedMatrix}
\alias{print.normalizedMatrix}
\title{
Print normalized matrix
}
\description{
Print normalized matrix
}
\usage{
\method{print}{normalizedMatrix}(x, ...)
}
\arguments{

  \item{x}{the normalized matrix returned by \code{\link{normalizeToMatrix}}}
  \item{...}{other arguments}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
