\name{rbind.normalizedMatrix}
\alias{rbind.normalizedMatrix}
\title{
Bind matrix by rows
}
\description{
Bind matrix by rows
}
\usage{
rbind.normalizedMatrix(..., deparse.level = 1)
}
\arguments{

  \item{...}{matrices}
  \item{deparse.level}{-deparse.level}

}
\value{
A \code{normalizedMatrix} class object.
}
\author{
z.gu@dkfz.de
}
\examples{
# There is no example
NULL

}
