\name{score_fun}
\alias{score_fun}
\title{
Score function for matrix rows

}
\description{
Score function for matrix rows

}
\usage{
score_fun(x1, x2, x3)}
\arguments{

  \item{x1}{a vector corresponding to upstream}
  \item{x2}{-x2}
  \item{x3}{-x3}
}
\details{
After obtaining a matrix from \code{\link{normalizeToMatrix}}, rows are ordered
in order to capture the features.

This returned value are summation of three values

}
\value{
a numeric value

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
