\name{show-dispatch}
\alias{show}
\title{
Method dispatch page for show}
\description{
Method dispatch page for \code{show}.}
\section{Dispatch}{
\code{show} can be dispatched on following classes:

\itemize{
\item \code{\link{show,CentralizedHeatmapList-method}}, \code{\link{CentralizedHeatmapList-class}} class method
\item \code{\link{show,CentralizedHeatmap-method}}, \code{\link{CentralizedHeatmap-class}} class method
}
}
