\name{copyAttr}
\alias{copyAttr}
\title{
Copy attributes to another object
}
\description{
Copy attributes to another object
}
\usage{
copyAttr(x, y)
}
\arguments{

  \item{x}{object 1}
  \item{y}{object 2}

}
\details{
The \code{\link{normalizeToMatrix}} object actually is a matrix but with more additional attributes attached.
This function is used to copy these new attributes when dealing with the matrix.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
NULL
}
