\name{draw-dispatch}
\alias{draw}
\title{
Method dispatch page for draw
}
\description{
Method dispatch page for \code{draw}.
}
\section{Dispatch}{
\code{draw} can be dispatched on following classes:

\itemize{
\item \code{\link{draw,EnrichedHeatmapList-method}}, \code{\link{EnrichedHeatmapList}} class method
\item \code{\link{draw,EnrichedHeatmap-method}}, \code{\link{EnrichedHeatmap}} class method
}

}
\section{Example}{
# no example
NULL
}
