\name{EnrichedHeatmap-class}
\docType{class}
\alias{EnrichedHeatmap-class}
\title{
Class for a single heatmap
}
\description{
Class for a single heatmap
}
\details{
The \code{\link{EnrichedHeatmap-class}} is inherited from \code{\link[ComplexHeatmap]{Heatmap-class}}.
}
\section{Methods}{
The \code{\link{EnrichedHeatmap-class}} provides following methods:

\itemize{
  \item \code{\link{EnrichedHeatmap}}: constructor method.
  \item \code{\link{draw,EnrichedHeatmap-method}}: draw a single heatmap.
}}
<<<<<<< HEAD
\seealso{
\code{\link{EnrichedHeatmapList-class}}
}
=======
>>>>>>> bioc/master
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
