\name{[.normalizeToMatrix}
\alias{[.normalizeToMatrix}
\title{
Subset normalized matrix by rows

}
\description{
Subset normalized matrix by rows

}
\usage{
\method{[}{normalizeToMatrix}(x, i, j)}
\arguments{

  \item{x}{the normalized matrix returned by \code{\link{normalizeToMatrix}}}
  \item{i}{row index}
  \item{j}{column index, disabled}
}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
