\name{print.normalizedMatrix}
\alias{print.normalizedMatrix}
\title{
Print the Normalized Matrix
}
\description{
Print the Normalized Matrix
}
\usage{
\method{print}{normalizedMatrix}(x, ...)
}
\arguments{

  \item{x}{The normalized matrix returned by \code{\link{normalizeToMatrix}}.}
  \item{...}{Other arguments.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
