\name{rbind.normalizedMatrix}
\alias{rbind.normalizedMatrix}
\title{
Bind Matrix by Rows
}
\description{
Bind Matrix by Rows
}
\usage{
\method{rbind}{normalizedMatrix}(..., deparse.level = 1)
}
\arguments{

  \item{...}{Matrices}
  \item{deparse.level}{Not used.}

}
\value{
A \code{normalizedMatrix} class object.
}
\author{
z.gu@dkfz.de
}
\examples{
# There is no example
NULL
}
