\name{show-dispatch}
\alias{show}
\title{
Method dispatch page for show
}
\description{
Method dispatch page for \code{show}.
}
\section{Dispatch}{
\code{show} can be dispatched on following classes:

\itemize{
\item \code{\link{show,EnrichedHeatmapList-method}}, \code{\link{EnrichedHeatmapList}} class method
\item \code{\link{show,EnrichedHeatmap-method}}, \code{\link{EnrichedHeatmap}} class method
}

}
\section{Example}{
# no example
NULL
}
